updates and bugfixes
A counter has been added. This stores the number of calls to archive.org in a file. Please consider donating $1 for every 5000 page requests to offset the operating costs of the Wayback Machine.

The ability to set a target date has been added. Normally, the Wayback Machine will return the capture closest to the midpoint between the earliest valid date matching the date_part and the latest date matching the date_part. This parameter allows setting a different target, while still allowing a broad range of results.

The handling of the flag changed (on the archive.org side), making the default no longer work on the newest releases of Matlab. The default has been changed from an asterisk to an empty char (this required a change in the parser).

A minor change to the tester was required to make it pass on Octave 7.1.

Several updates to dependencies, including improvements in the fuzzy matching of Name,Value parameters.